+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add argument `orderRegion` and `orderReads = "region"` to read-level plotting functions to sort reads by average modification probabilities in a specified window
 * Add argument `minCoveredFraction` to `filterReads()` and `plotRegion()`
 * Add argument `fillColors` to read-level plotting functions to control colors of modification probabilities
+* Add arguments `clustDist` and `windowWidth` to read-level plotting functions for more fine-grained control of clustering
 
 # footprintR 0.3.4