+++ NEWS.md
@@ -1,3 +1,7 @@
+# footprintR 0.3.6
+
+* Filter out reads without modified-base calls if `maxFracLowConf` or `maxEntropy` filters are set in `filterReads` or `filterReadsBam`
+
 # footprintR 0.3.5
 
 * Add argument `orderRegion` and `orderReads = "region"` to read-level plotting functions to sort reads by average modification probabilities in a specified window